Images appear in grid
still working on setting  attributes to each image and making the images
disappear when they are clicked on in the new grid .
to do list:
Assign points to each image.
create score : each image will have 100 points, when the image coincides
with the image from the first grid. 100 points are awarded and vice
versa.
timer has to be adjusted to each grid. there should be a total time that
counts down that should always be on the second grid. The timer should
be paused when grid 1 is up and resume when grid 2 reappears.